postgresql-dump-script/requirements.txt
gzippy
delegator.py